feat(connection): add TCP connection implementation This commit adds a new implementation of the Connection interface for TCP connections. The TcpConnection class implements the virtual functions defined in the Connection interface. It provides functionality to connect to a server, start a server, accept a client, disconnect, send data, and receive data.
feat(udp.cpp): add implementation for UdpConnection class
refactor(networkssy.cpp): remove unused code and move socket class to udp.cpp
The UdpConnection class has been added with implementation for its methods. The class provides functionality for sending and receiving data over a UDP connection. The socket class has been moved to udp.cpp and the tcp_connection and tcp_socket classes have been removed as they are not used in the code. The fn() function has also been removed as it is not used.

feat(socket): add support for TCP and UDP sockets
This commit adds support for TCP and UDP sockets. The Socket class is a base class that provides basic functionality for creating and closing sockets. The TcpSocket class extends the Socket class and provides functionality for connecting to a remote host, listening for incoming connections, and accepting incoming connections. The UdpSocket class also extends the Socket class and provides functionality for sending and receiving data over UDP.